requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.3.4
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.11.2
dj-database-url==0.5.0
Django==3.2.2
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==2.10
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==2.11.3
joblib==1.1.0
jsonschema==4.4.0
lru-dict==1.1.7
MarkupSafe==1.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.1
packaging==20.9
pandas==1.3.5
parsimonious==0.8.1
protobuf==3.20.1
psycopg2==2.8.6
pycryptodome==3.14.1
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
rlp==2.0.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==3.0.0
toolz==0.11.2
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.4
varint==1.0.2
web3==5.29.0
websockets==9.1
whitenoise==5.2.0
yarl==1.7.2
